WIP: install basic system

+++ util/clean.sh
@@ -9,6 +9,8 @@ fi
 
 losetup -D
 rm -rfv $BUILD_DIR
-rm -rfv $DOWNLOAD_DIR
+if [[ $KEEP_DOWNLOAD_FILES -eq 0 ]]; then
+    rm -rfv $DOWNLOAD_DIR
+fi
 rm -rfv $BACKUP_DIR
 userdel $LFS_USER
